Refactor configuration variables and constant file stuff into its own class
