+++ lib/intercode/import/intercode1.rb
@@ -25,6 +25,7 @@ module Intercode
         "#{severity_rep} #{col.dark(col.yellow(time.strftime('%H:%M:%S.%L')))} - #{msg}\n"
       end
 
+      autoload :Configuration
       autoload :DateHelpers
       autoload :EmbeddedPdfPage
       autoload :FormImporter
